Fixed the FIXME (#30)
Added an or to the if statement. If the version was set as blank, it would think it has a setting,
thus making this not work properly. Removed the set_initial_value as this was making the
default value whatever it was set to (1.0.0), which was making Godot remove it from the
project.godot file because of the way that Godot handles the default value(?).

Also, removed the second half of the fixme as it was already in the main debug script.

Co-authored-by: Hugo Locurcio <hugo.locurcio@hugo.pro>